# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PUCPR Projeto Final - Webservices e MBaaS ",
        "description": "API criada como projeto final da matéria de Webservices e MBaaS - por Lucas Cândido dos Santos",
        "contact": {
            "name": "Lucas Cândido dos Santos"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/item-pedidos": {
            "put": {
                "tags": [
                    "itens"
                ],
                "summary": "Edita um item de um pedido existente",
                "description": "Edita um item de um pedido existente",
                "operationId": "6001a4350fcd97e09ed71992589982fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Informe o ID do item desejado",
                        "required": true,
                        "example": "123"
                    },
                    {
                        "name": "produto_id",
                        "in": "query",
                        "description": "ID do produto",
                        "required": false
                    },
                    {
                        "name": "quantidade",
                        "in": "query",
                        "description": "Quantiade comprada do item",
                        "required": false
                    },
                    {
                        "name": "unitario",
                        "in": "query",
                        "description": "Valor unitário do item",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado. Apenas administradores podem cadastrar produtos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "itens"
                ],
                "summary": "Adiciona um item a um pedido existente",
                "description": "Adiciona um item a um pedido existente",
                "operationId": "acd77c2f7774d381b0bd0ae37cc6964a",
                "parameters": [
                    {
                        "name": "pedido_id",
                        "in": "query",
                        "description": "ID do pedido a que o item pertence",
                        "required": true
                    },
                    {
                        "name": "produto_id",
                        "in": "query",
                        "description": "ID do produto",
                        "required": true
                    },
                    {
                        "name": "quantidade",
                        "in": "query",
                        "description": "Quantiade comprada do item",
                        "required": true
                    },
                    {
                        "name": "unitario",
                        "in": "query",
                        "description": "Valor unitário do item",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado. Apenas administradores podem cadastrar produtos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/item-pedidos/{id}": {
            "get": {
                "tags": [
                    "itens"
                ],
                "summary": "Retorna um produto do pedido pelo ID",
                "description": "Retorna um produto do pedido pelo ID",
                "operationId": "e3c9d276ae5041d11fcca46ee4678906",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Informe o ID do item desejado",
                        "required": true,
                        "example": "123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "itens"
                ],
                "summary": "Remove um produto do pedido pelo ID",
                "description": "Remove um produto do pedido pelo ID",
                "operationId": "e7044bd56104f6a50515ab3d51acc2c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Informe o ID do item desejado",
                        "required": true,
                        "example": "123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}